feat: create VolumeBackup component
